test(review): co-locate guarded raw repair coverage
